# UgramApiSequilize/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Ugram Api",
    "description": "Api for Ugram Project",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users in system",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Get Specific page",
            "type": "number"
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "description": "number of items per page",
            "type": "number"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Value of your generic search",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Request OK",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          }
        }
      }
    },
...
    "/pictures": {
      "get": {
        "tags": [
          "Pictures"
        ],
        "summary": "Get all pictures",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Get Specific picture page",
            "type": "number"
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "description": "number of items per page",
            "type": "number"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Value of your generic search",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Request OK",
            "schema": {
              "$ref": "#/definitions/Pictures"
            }
          }
        }
      }
    },
...
    "/oauth/token": {
      "post": {
        "tags": [
          "Login - Register OAuth2"
        ],
        "summary": "Login - Get new token",
        "consumes": [
          "application/x-www-form-urlencoded"
        ]
        ,
        "parameters": [ 
          {
              "name": "username",
              "in": "formData",
              "required": true,
              "description": "username",
              "type": "string"
            },
            {
              "name": "password",
              "in": "formData",
              "required": true,
              "description": "password",
              "type": "string"
            },
            {
              "name": "client_id",
              "in": "formData",
              "required": true,
              "description": "client id oauth",
              "type": "string",
              "value": "abc123"
            },
            {
              "name": "client_secret",
              "in": "formData",
              "required": true,
              "description": "client secret oauth",
              "type": "string",
              "value": "ssh-password"
            },
            {
              "name": "grant_type",
              "in": "formData",
              "required": true,
              "description": "Value of your generic search",
              "type": "string",
              "value": "password"
            }
          ],
        "responses": {
          "200": {
            "description": "Request token",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          } ,
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "501": {
            "description": "Bad grand_type"
          }
        },
        "security": [
            {
                "login": []
            }
        ]
      }
    },
    "/oauth/register": {
      "post": {
        "tags": [
          "Login - Register OAuth2"
        ],
        "summary": "register new client",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
              "name": "username",
              "in": "formData",
              "required": true,
              "description": "username",
              "type": "string"
            },
            {
              "name": "password",
              "in": "formData",
              "required": true,
              "description": "password",
              "type": "string"
            },
            {
              "name": "client_id",
              "in": "formData",
              "required": true,
              "description": "client id oauth",
              "type": "string",
              "value": "abc123"
            },
            {
              "name": "client_secret",
              "in": "formData",
              "required": true,
              "description": "client secret oauth",
              "type": "string",
              "value": "ssh-password"
            },
            {
              "name": "grant_type",
              "in": "formData",
              "required": true,
              "description": "Value of your generic search",
              "type": "string",
              "value": "password"
            }
          ],
        "responses": {
          "200": {
            "description": "Register new User"
          } ,
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
            {
                "login": []
            }
        ]
      }
    }
  },
  "definitions": {
    "User": {
      "required": [
        "email",
        "id"
      ],
      "properties": {
        "id": {
          "type": "string",
          "uniqueItems": true
        },
        "email": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "number"
        },
        "pictureUrl": {
          "type": "string"
        },
        "registrationDate": {
          "type": "number"
        }
      }
    },
    "Users": {
      "properties": {
        "totalEntries": {
          "type": "number"
        },
        "totalPages": {
          "type": "number"
        },
        "items":  {
            "type": "array",
            "items": {
              "$ref": "#/definitions/User"  
            }
          }
      }
    },
    "Like": {
      "properties": {
        "value": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      }
    },
    "Picture": {
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "number",
          "uniqueItems": true
        },
        "createdDate": {
          "type": "number"
        },
        "description": {
          "type": "string"
        },
        "mentions": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "url": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "likes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Like"
          }
        },
        "nbLike": {
          "type": "number"
        }
      }
    },
...
    "Pictures": {
      "properties": {
        "totalEntries": {
          "type": "number"
        },
        "totalPages": {
          "type": "number"
        },
        "items":  {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Picture"  
            }
          }
      }
    },
    "Token": {
      "properties": {
        "access_token": {
          "type": "string"
        },
        "id": {
          "type": "number"
        },
        "token_type" : {
          "type": "string"
        }
      }
    }
  }
}